Moving utility method from sqldb to mysql.
